Resolve implicit type conversions on 64-bit archs

+++ Tropos/Categories/UIScrollView+TRReactiveCocoa.m
@@ -14,7 +14,7 @@
 - (RACSignal *)verticalAmountScrolledSignal
 {
     return [[RACObserve(self, contentOffset) map:^id(NSValue *contentOffset) {
-        CGFloat yOffset = floor(contentOffset.CGPointValue.y);
+        CGFloat yOffset = (CGFloat)floor(contentOffset.CGPointValue.y);
         return @(fabs(yOffset));
     }]
     distinctUntilChanged]
